Ditch `npm-run-all`
