+++ internals/get-package-json/package.json
@@ -8,9 +8,7 @@
     "dist"
   ],
   "scripts": {
-    "build": "run-p build:*",
-    "build:js": "node ../../scripts/esbuild.mjs",
-    "build:types": "tsc",
+    "build": "node ../../utils/build.mjs",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail",
     "test-unit": "pnpm test tests/unit"
   },
