+++ package.json
@@ -29,7 +29,6 @@
     "lint-staged": "9.2.5",
     "node-fetch": "2.6.7",
     "npm-package-arg": "6.1.0",
-    "npm-run-all": "4.1.5",
     "prettier": "2.6.2",
     "source-map-support": "0.5.12",
     "ts-eager": "2.0.2",
