+++ packages/build-utils/build.mjs
@@ -1,0 +1,6 @@
+import { tsc, esbuild } from '../../utils/build.mjs';
+
+await Promise.all([
+    tsc(),
+    esbuild().then(() => esbuild({ bundle: true }))
+]);
