+++ packages/build-utils/package.json
@@ -11,11 +11,7 @@
     "directory": "packages/now-build-utils"
   },
   "scripts": {
-    "build": "run-p build:js build:types",
-    "build:js": "run-s build:single build:bundle",
-    "build:single": "node ../../scripts/esbuild.mjs",
-    "build:bundle": "node ../../scripts/esbuild.mjs --bundle",
-    "build:types": "tsc",
+    "build": "node build.mjs",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail",
     "test-unit": "pnpm test test/unit.*test.*",
     "test-e2e": "pnpm test test/integration.test.ts"
