+++ packages/client/package.json
@@ -14,9 +14,7 @@
     "directory": "packages/client"
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:js": "node ../../scripts/esbuild.mjs",
-    "build:types": "tsc",
+    "build": "node ../../utils/build.mjs",
     "test-e2e": "pnpm test tests/create-deployment.test.ts tests/create-legacy-deployment.test.ts tests/paths.test.ts",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail",
     "test-unit": "pnpm test tests/unit.*test.*"
