+++ packages/error-utils/package.json
@@ -13,8 +13,7 @@
     "directory": "packages/error-utils"
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:js": "node ../../scripts/esbuild.mjs",
+    "build": "node ../../utils/build.mjs",
     "build:types": "tsc",
     "test": "jest --reporters=default --reporters=jest-junit --coverage --env node --verbose",
     "test-unit": "pnpm test"
