+++ packages/frameworks/package.json
@@ -13,9 +13,7 @@
   },
   "license": "Apache-2.0",
   "scripts": {
-    "build": "run-p build:*",
-    "build:src": "node ../../scripts/esbuild.mjs",
-    "build:types": "tsc",
+    "build": "node ../../utils/build.mjs",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail",
     "test-unit": "pnpm test"
   },
