+++ packages/remix/package.json
@@ -10,9 +10,7 @@
     "directory": "packages/remix"
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:js": "node ../../scripts/esbuild.mjs",
-    "build:types": "tsc",
+    "build": "node ../../utils/build.mjs",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --bail --runInBand",
     "test-unit": "pnpm test test/unit.*test.*",
     "test-e2e": "pnpm test test/integration.test.ts"
