+++ packages/static-config/package.json
@@ -9,9 +9,7 @@
     "directory": "packages/static-config"
   },
   "scripts": {
-    "build": "run-p build:*",
-    "build:js": "node ../../scripts/esbuild.mjs",
-    "build:types": "tsc",
+    "build": "node ../../utils/build.mjs",
     "test-unit": "pnpm test",
     "test": "jest --reporters=default --reporters=jest-junit --env node --verbose --runInBand --bail"
   },
